requirements.txt
asgiref==3.5.2
Django==4.0.6
django-phonenumber-field==6.3.0
django-restframework==0.0.1
djangorestframework==3.13.1
phonenumbers==8.12.51
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
